orpo trainer: add a low VRAM threshold for get_batch_logps()
This method will cause pretty huge spikes in VRAM for high token samples. This pull request lets the user pick a context length (hinted to in the crash when running out of VRAM) which transfers the tensors to CPU and performs the operation there instead, at a small performance cost.