Update profiling.py: fix scoping problems for wandb and mlflow
In the current version, wandb and mlflow goes out of scope once we exit the if statement. 

When we run train() on GRPOTrainer with report_to=['wandb'] parameter specified in GRPOConfig, the program fails in line 64 (if "wandb" in trainer.args.report_to and wandb.run is not None and trainer.accelerator.is_main_process) with NameError.

This is the error message with the current version:
"NameError: name 'wandb' is not defined"

After the patch, the run is successful.

I anticipate the same issue to exist for mlflow so I applied the same fix on that as well.