CPOTrainer - Incorrect handling of different length chosen/rejected prompts.
on line 495, zip() function has a flag `strict=True`, and only after that we calculate the num_diff_len, which we can not reach in case the chosen and rejected prompt lengths differ by 1 (as zip raises an exception). Therefore, we should take only the first `prompt_len_input_ids` common parts of the prompts, and find length differences afterwards.